+++ demisto_sdk/commands/common/constants.py
@@ -34,6 +34,12 @@ DOCUMENTATION_DIR = 'Documentation'
 JOBS_DIR = 'Jobs'
 PRE_PROCESS_RULES_DIR = 'PreProcessRules'
 LISTS_DIR = 'Lists'
+PARSING_RULES_DIR = 'ParsingRules'
+MODELING_RULES_DIR = 'ModelingRules'
+CORRELATION_RULES_DIR = 'CorrelationRules'
+XSIAM_DASHBOARDS_DIR = 'XSIAMDashboards'
+XSIAM_REPORTS_DIR = 'XSIAMReports'
+TRIGGER_DIR = 'Triggers'
 
 SCRIPT = 'script'
 AUTOMATION = 'automation'
@@ -100,6 +106,7 @@ class FileType(Enum):
     AUTHOR_IMAGE = 'author_image'
     DOC_IMAGE = 'doc_image'
     PYTHON_FILE = 'pythonfile'
+    XIF_FILE = 'xiffile'
     JAVASCRIPT_FILE = 'javascriptfile'
     POWERSHELL_FILE = 'powershellfile'
     CONF_JSON = 'confjson'
@@ -116,6 +123,12 @@ class FileType(Enum):
     PRE_PROCESS_RULES = 'pre-process-rule'
     LISTS = 'list'
     JOB = 'job'
+    PARSING_RULE = 'parsingrule'
+    MODELING_RULE = 'modelingrule'
+    CORRELATION_RULE = 'correlationrule'
+    XSIAM_DASHBOARD = 'xsiamdashboard'
+    XSIAM_REPORT = 'xsiamreport'
+    TRIGGER = 'trigger'
 
 
 RN_HEADER_BY_FILE_TYPE = {
@@ -171,6 +184,15 @@ ENTITY_TYPE_TO_DIR = {
     FileType.LISTS.value: LISTS_DIR,
     FileType.JOB.value: JOBS_DIR
 }
+
+SIEM_ONLY_ENTITIES = [
+    FileType.PARSING_RULE.value,
+    FileType.MODELING_RULE.value,
+    FileType.CORRELATION_RULE.value,
+    FileType.XSIAM_DASHBOARD.value,
+    FileType.XSIAM_REPORT.value,
+    FileType.TRIGGER.value
+]
 
 CONTENT_FILE_ENDINGS = ['py', 'yml', 'png', 'json', 'md']